requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
PyMuPDF>=1.23.0
openpyxl>=3.1.0
python-docx>=1.0.0
paddlepaddle>=3.0.0
paddleocr>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
